sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dekode"
version = "0.1.0"
description = "Python SDK for the Dekode Agent Protocol"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.62.0",
    "protobuf>=4.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "grpcio-tools>=1.62.0",
]
anthropic = [
    "anthropic>=0.40.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dekode*"]

[tool.setuptools.package-data]
dekode = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires running dk-server (deselect with '-m not integration')",
]
